tools/requirements.txt
mem0ai
tavily-python
pyautogui
mss
litellm
python-dotenv
